# Conflicts:
#	marketing-cs/src/main/java/com/br/marketing/service/Impl/MarketingCustomerServiceImpl.java
#	marketing-cs/src/main/java/com/br/marketing/service/Impl/VariableDicServiceImpl.java
#	marketing-cs/src/main/java/com/br/marketing/service/MarketingCustomerService.java
#	marketing-cs/src/main/java/com/br/marketing/vo/MarketingCustomerListVO.java
#	marketing-inner-api/src/main/java/com/br/marketing/innerapi/controller/CustomerController.java

+++ marketing-cs/src/main/java/com/br/marketing/mapper/SyncConfigMapperBase.java
@@ -11,7 +11,7 @@ public interface SyncConfigMapperBase {
 
     int deleteByExample(SyncConfigExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SyncConfig record);
 
@@ -19,7 +19,7 @@ public interface SyncConfigMapperBase {
 
     List<SyncConfig> selectByExample(SyncConfigExample example);
 
-    SyncConfig selectByPrimaryKey(Integer id);
+    SyncConfig selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") SyncConfig record, @Param("example") SyncConfigExample example);
